requirements.txt
absl-py==0.8.1
asn1crypto==0.24.0
astor==0.8.0
backports.weakref==1.0.post1
certifi==2019.11.28
chardet==3.0.4
configparser==3.5.0b2
cryptography==2.6.1
docutils==0.15.2
entrypoints==0.3
enum34==1.1.6
funcsigs==1.0.2
gast==0.3.2
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
idna==2.8
ipaddress==1.0.17
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==18.0.1
keyrings.alt==3.1.1
Kivy-Garden==0.1.4
Markdown==3.1.1
mock==3.0.5
numpy==1.16.5
protobuf==3.11.1
pycrypto==2.6.1
Pygments==2.5.2
pyxdg==0.25
requests==2.22.0
SecretStorage==2.3.1
six==1.12.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
urllib3==1.25.7
virtualenv==16.7.8
Werkzeug==0.16.0
wrapt==1.11.2
